door ctrl: Rework the event handling to prevent potential races
Currently the door controller state is changed from timer handlers.
This is a problem because they run from interrupt handlers so the
state might change while we handle an event.

To prevent this we now generate events from the timeout handler and
then handle the state change from the event handler. To spare some
memory we use the same handler as for the wiegand events, so the events
are added with the wiegand reader as source and not the whole
controller.

Signed-off-by: Alban Bedel <albeu@free.fr>