firmware: Add an UART based control protocol
The control protocol allow to query the device config, get/set the
doors config and the access records. The commands handling is splitted
from the transport layer to allow later adding other transport such as
USB. The ctrl-cmd header is also splitted to allow reusing the data
type in host tools that will use this protocol.

Signed-off-by: Alban Bedel <albeu@free.fr>