eeprom: Add an invalid flag to make sure a blank eeprom isn't valid
When the firmware is installed on a device with a blank eeprom we
don't want records fulls of 1's to be considered valid. Without this
flag such a blank record gives access to all doors to any card^pin
combination that result in 0xFFFFFFFF. Such a sequence could easily
be generated using a properly crafted card/pin combination, or
by directly toggling the D0 and D1 lines.

Signed-off-by: Alban Bedel <albeu@free.fr>