ctrl-cmd: Add support for sending events
Events might be sent at any time, however both events and command
reply are sent from the main execution context. As such if the sending
buffer is already in use we can just wait for the sending to finish.
With the current maximum message size we have a worse case of
(2 + 16) * 2 + 1 = 37 bytes which take (37 * 10) / 38400 = 9.6 ms.
So in practice we should never have to wait for more than a few ms.

Furthermore a command might be received while an event is sent, so we
now need 2 buffer, so we have to remove the union that was used to
save some memory.

While at it also set the status variable volatile as it is read from
both the main context and interrupt context.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>