firmware: gpio: Make the polarity part of the GPIO ID
Currently we need to have a 'low_active' field next to most gpio, this
take some RAM and code as each user reimplement xor'ing the low_active
field with the GPIO value.

To fix this we now make the polarity part of the GPIO ID, in this way
the GPIO code can internally take care of this. That reduce the RAM
usage as we get rid of all the low_active fields and code size as we
spare various extra xor.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>