firmware: boards: nano v3: Set the config for carrier rev 0
We now finally have the "real" carrier with true PCB, however these
have a slightly different pinning than the prototypes build on
perfboards. To handle this correctly we would need to support
different board revisions. Not a big deal but as we only used arduino
nano v2 on the prototypes, and the new devices all get arduino nano v3,
we just use this to differentiate the hardware versions for now.

Signed-off-by: Alban Bedel <albeu@free.fr>