electronic: schematics: Add a jumper to select the input voltage
To support both the step-up and step-down config we need to connect
either the +12V or the +24V with the input voltage. This would mean
adding a ratline on the PCB, which is not very nice. Instead we add a
jumper footprint where one can easily solder a wire, along with the
label it make it clearly visible which input voltage the board expect.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>