client: db: Add `on delete` and `on update` relation constraints
Add constraints to all the foreign keys to make it easier to work with
the DB. The constraints take care to automatically update or delete
records in all the secondary tables. All tables with foreign keys now
delete the relevant records when the parent entry is delete, except
for the Doors table that set the ControllerID to null if the
controller is deleted.

This allow to delete user, group, door and controller without having
to first manually delete all related records.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>