client: admin: Add commands to edit a user groups membership
Currently group membership can only be manipulated from the group
command. However it is more common for an admin to change the
memberships of a single user than to add/remove many users to/from a
group. To make this simpler add the 'add-to-group' and
'remove-from-group' sub-command to the user command.

Signed-off-by: Alban Bedel <albeu@free.fr>