eeprom: Fix the error code used in eeprom_get/set_access_record
With the current implementation the only error that can happen is that
the index is out of bound. However returning -EINVAL in this case is a
bit too generic as the caller can't known the size of the storage.
Return -ENOENT instead to allow caller to stop iterating while not
ignoring other errors that might happen with another implementation.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>