firmware: Add a new API for deferred execution
The event-queue API turned out to be a bit clumsy to use. The multi
receiver capability has never been used and it use too much memory on
the event handler side.

The new work-queue API attempts to solve the same problem in a simpler
way. Instead of events it allow to queue the execution of a function.
To make it easy to use with the existing code the queued function use
the same arguments as the event handlers.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>