firmware: Rework the EEPROM storage to prepare for OTP pin
To allow supporting OTP pins we need to rework the EEPROM storage as
the trick to XOR the card and pin is not possible anymore. The new
scheme doesn't need more space for the common case of only card or
pin and allow for 3 type of PIN. Instead of xor'ing both data they are
stored one after the other on the EEPROM. In memory the new records
provide both the card and pin data for ease of use.

A backward compatibility layer is included to avoid the need to change
the API users right away. There is only limited support for pin+card
through the compatibility layer, basically read-only, as the xor can't
be undone.

Signed-off-by: Alban Bedel <albeu@free.fr>