client: Improve events handling
Currently event received before a command reply would just be ignored.
Add a queue to store them and a function to read them that start by
emptying the queue, this way we don't drop any event. Then add a
parsing function so we can display them in a human readable form.
Finally improve the `show_events` function to skip over timeouts and
stop on keyboard interrupt.

Signed-off-by: Alban Bedel <albeu@free.fr>