client: DB: Add server side OTP support
Implementing OTP support in the firmware is major change that is still
not finished. As a workaround we instead implement this feature in the
database. For the rest of the software stack the OTP PIN are just
normal PIN that are generated by the database. A big down side of this
is the strain on the controllers EEPROM, so timeout shorter than 8 hours
shouldn't be used. Also the controllers will get out of sync if they
are offline for too long, or if the update cron job fails to run. On
the positive side the controllers don't need an RTC module.

Signed-off-by: Alban Bedel <albeu@free.fr>