client: Add support for the new access records
The new access records add support for HOTP and TOTP. All access
record related methods in the AVRDoorCtrl class can now return either
the old or new format. When possible the native transport commands for
the requested format are used. If they are not supported a fallback is
used with the other command and the result is converted.

Signed-off-by: Alban Bedel <albeu@free.fr>