firmware: Add a controller config to store the OTP root key
A root key is needed to derive the OTP secrets, it is stored in EEPROM
before the door configuration. This loose some access record space so
the opportunity is used to also remove unused fields in the door config.
Two control commands are added to allow getting and setting the
controller config. Unless build with debug the getter always return an
empty key. Similarly the setter only overwrite the key if a new one is
provided.

Signed-off-by: Alban Bedel <albeu@free.fr>