firmware: Add OTP PIN support
Add an SHA1 implementation with HMAC, that is then used to implement
HOTP and TOTP. The OTP secrets are derived using HKDF with the
controller key and a 10 bits OTP "key ID", allowing up to 1024 unique
keys when only a PIN is used. When a card is used the card ID is also
passed to the key derivation to ensure each user get a unique key.

Signed-off-by: Alban Bedel <albeu@free.fr>