firmware: Avoid hanging when failling to send a command reply
When `ctrl_transport_reply` fails to write out its paylod, for example
because it is too large, it would switch to the
`UART_CTRL_TRANSPORT_SYNC` state which prevent sending a new error
reply. Update the condition to switch back to the
`UART_CTRL_TRANSPORT_WAIT_FOR_REPLY` state instead to give the caller
the chance to send an error reply.

Signed-off-by: Alban Bedel <albeu@free.fr>